bugfixes & improv. win/zipto7z.ps1

+++ win/zipto7z.ps1
@@ -2,15 +2,19 @@
 # DEPENDENCIES: requires 7z.exe to be in your $PATH
 Foreach ($zip in $(Get-ChildItem -Filter *.zip)) {
 	Set-Variable -Name name -Value $zip.toString().TrimEnd(".zip")
-	Write-Output $name
-	Expand-Archive -Path $zip
+	If (Test-Path "$name.7z") { continue }
+
+	Write-Output "$name"
+	Expand-Archive -Path "$zip"
 	
-	Set-Directory $name
-	7z a -mx9 $name
+	Set-Location "$name"
+	7z a -mx9 "$name"
 	
 	Move-Item "$name.7z" ..
 	Set-Location ..
 	
-	Remove-Item $zip
-	Remove-Item $name -Recurse
+	If (Test-Path "$name.7z") {
+		Remove-Item "$zip"
+		Remove-Item "$name" -Recurse
+	}
 }
